studio: split viewer.rs into api/paint_jobs + topical test modules
