+++ crates/mogen-studio/src/viewer.rs
@@ -1,4 +1,5 @@
 mod anim;
+mod api;
 mod camera;
 mod cinema;
 mod colliders_gl;
@@ -10,17 +11,17 @@ mod grid_gl;
 mod imposter_gl;
 mod lights;
 mod lights_gl;
+mod paint_jobs;
 mod renderer;
 mod shaders;
 pub mod shadows;
 pub(crate) mod state;
 
-use std::path::Path;
 use std::sync::{Arc, Mutex};
 
 use eframe::egui;
 use glam::Mat4;
-use mogen_core::{NodeId, SceneGraph};
+use mogen_core::NodeId;
 
 pub use camera::{CameraSnapshot, OrbitCamera};
 pub use environment::Environment;
@@ -33,13 +34,10 @@ pub use state::{
     ImposterRequest, PendingEdit, SelectionPath,
 };
 
-use crate::preview_shader::PreviewShader;
-
 use renderer::Renderer;
 use state::{
-    aspect_for, begin_gizmo_drag, commit_gizmo_drag, find_deepest_node_at_offset,
-    find_use_at_offset, gizmo_handles_supported, node_path, replace_selection,
-    replace_selection_cycling, resolve_node_path, toggle_selection, update_gizmo_drag,
+    aspect_for, begin_gizmo_drag, commit_gizmo_drag, gizmo_handles_supported,
+    replace_selection, replace_selection_cycling, toggle_selection, update_gizmo_drag,
     ViewerState,
 };
 
@@ -62,546 +60,10 @@ impl Viewer {
         })
     }
 
-    pub fn set_scene(&self, scene: Arc<SceneGraph>, base_dir: Option<&Path>, fit_camera: bool) {
-        let mut st = self.state.lock().unwrap();
-        let viewer_was_empty = st.scene.is_none();
-        st.base_dir = base_dir.map(|p| p.to_path_buf());
-        // Re-resolve every saved path against the new scene; drop any that
-        // no longer resolve (a node deleted by the latest edit just falls
-        // out of selection silently).
-        let mut new_selected = Vec::with_capacity(st.selected_paths.len());
-        let mut new_paths = Vec::with_capacity(st.selected_paths.len());
-        for path in &st.selected_paths {
-            if let Some(id) = resolve_node_path(&scene, path) {
-                new_selected.push(id);
-                new_paths.push(path.clone());
-            }
-        }
-        st.selected = new_selected;
-        st.selected_paths = new_paths;
-        st.gizmo_drag = None;
-        // NodeIds may have been renumbered by the recompile, so the
-        // recorded leaf in `pick_cycle` no longer refers to a stable
-        // node. Reset the cycle so the next click starts fresh.
-        st.pick_cycle = None;
-
-        let prev_active: Vec<String> = match &st.scene {
-            Some(prev) => prev
-                .clips
-                .iter()
-                .enumerate()
-                .filter_map(|(i, c)| {
-                    (*st.clip_active.get(i).unwrap_or(&false)).then(|| c.name.clone())
-                })
-                .collect(),
-            None => Vec::new(),
-        };
-        // Only treat the incoming scene as a recompile of the current scene
-        // (and preserve user toggles) when at least one clip name is shared
-        // with the previous scene. A swap to an unrelated scene — including
-        // the cascade of `set_scene` calls when the studio restores multiple
-        // tabs at startup — has no overlap, so default its clips to active
-        // instead of leaving them all off and hiding the user's animations.
-        let prev_clip_names: std::collections::HashSet<&str> = match &st.scene {
-            Some(prev) => prev.clips.iter().map(|c| c.name.as_str()).collect(),
-            None => std::collections::HashSet::new(),
-        };
-        let names_overlap = scene
-            .clips
-            .iter()
-            .any(|c| prev_clip_names.contains(c.name.as_str()));
-
-        let mut clip_active: Vec<bool> = vec![false; scene.clips.len()];
-        if viewer_was_empty || (!names_overlap && !scene.clips.is_empty()) {
-            for a in &mut clip_active {
-                *a = true;
-            }
-            st.anim_playing = !scene.clips.is_empty();
-        } else {
-            for (i, clip) in scene.clips.iter().enumerate() {
-                if prev_active.iter().any(|n| n == &clip.name) {
-                    clip_active[i] = true;
-                }
-            }
-        }
-        st.clip_active = clip_active;
-        st.anim_times = vec![0.0; scene.clips.len()];
-        st.scene = Some(scene);
-        // Single flatten via rebuild_mesh — vertex positions are rest-pose
-        // (animation is applied by the joint-palette uniform in the shader),
-        // so the resulting mesh AABB is the static framing bound. Pulling
-        // it from `st.mesh` here avoids a redundant flatten() purely for
-        // the camera-fit numbers.
-        st.rebuild_mesh();
-        st.static_center = st.mesh.center;
-        st.static_radius = st.mesh.radius;
-        // Camera fit is caller-driven: the App knows whether this is the first
-        // time a given file's scene is being shown. Subsequent compiles
-        // (gizmo release, keystroke debounce) pass `false` so `camera.target` /
-        // `camera.fit_distance` stay put — recentering on every compile would
-        // make it look like the user's edit was rejected. First-render uses
-        // the same helper as the Frame button so a freshly-loaded model
-        // composes identically to pressing Frame (yaw/pitch reset included).
-        if fit_camera {
-            Self::frame_camera_to_static(&mut st);
-        }
-    }
-
-    pub fn clear(&self) {
-        let mut st = self.state.lock().unwrap();
-        st.mesh = FlatMesh::default();
-        st.mesh_dirty = true;
-        st.scene = None;
-        st.base_dir = None;
-        st.clip_active.clear();
-        st.anim_times.clear();
-        st.selected.clear();
-        st.selected_paths.clear();
-        st.gizmo_drag = None;
-        st.pick_cycle = None;
-    }
-
-    /// True when no scene is loaded yet (fresh launch with an empty buffer).
-    /// Used by the viewport overlay to render a help message.
-    pub fn has_scene(&self) -> bool {
-        self.state.lock().unwrap().scene.is_some()
-    }
-
-    /// Replace the selection with a single node (or clear it). Equivalent
-    /// to a plain click in the viewport.
-    pub fn set_primary_selection(&self, id: Option<NodeId>) {
-        let mut st = self.state.lock().unwrap();
-        st.selected.clear();
-        st.selected_paths.clear();
-        if let Some(n) = id {
-            st.selected.push(n);
-            if let Some(path) = st.scene.as_ref().and_then(|s| node_path(s, n)) {
-                st.selected_paths.push(path);
-            }
-        }
-        st.gizmo_drag = None;
-        st.pick_cycle = None;
-    }
-
-    /// Most-recently-selected node — the one the gizmo, inspector, and
-    /// caret-jump follow. `None` when no node is selected.
-    pub fn primary_selection(&self) -> Option<NodeId> {
-        self.state.lock().unwrap().selected.last().copied()
-    }
-
-    /// Full selection set in click order; the last entry is the primary.
-    pub fn all_selected(&self) -> Vec<NodeId> {
-        self.state.lock().unwrap().selected.clone()
-    }
-
-    /// All stable name-paths in click order, parallel to `all_selected()`.
-    /// Used by the undo stack to capture / restore the full selection across
-    /// recompiles when raw `NodeId` indices may have shifted.
-    pub fn all_selected_paths(&self) -> Vec<SelectionPath> {
-        self.state.lock().unwrap().selected_paths.clone()
-    }
-
-    /// Reverse of the viewport-pick → caret jump: update the primary
-    /// selection to whichever user-authored node's `source_span` contains
-    /// `byte_offset`, so clicking in the code editor highlights the matching
-    /// 3D node. Does **not** queue a `pending_caret` — the caret is already
-    /// where the user moved it, and writing one back would re-trigger this
-    /// path next frame and pingpong forever. Returns `true` when the
-    /// selection actually changed; callers can short-circuit when nothing
-    /// moved. `None` from the offset lookup (caret in a comment, blank line,
-    /// or otherwise outside any span) preserves the existing selection.
-    pub fn select_node_at_source_offset(&self, byte_offset: usize, source: &str) -> bool {
-        let mut st = self.state.lock().unwrap();
-        let Some(scene) = st.scene.clone() else {
-            return false;
-        };
-        // A click on a `use "X" (...)` line maps to the imported root rather
-        // than whatever container would otherwise win the deepest-span tiebreak
-        // (typically `scene`, since imported nodes' spans live in another
-        // file and are skipped). Try this first so the editor click lands a
-        // meaningful viewport selection.
-        let target = find_use_at_offset(&scene, source, byte_offset)
-            .or_else(|| find_deepest_node_at_offset(&scene, byte_offset));
-        let Some(target) = target else {
-            return false;
-        };
-        if st.selected.last().copied() == Some(target) {
-            return false;
-        }
-        st.selected.clear();
-        st.selected_paths.clear();
-        st.selected.push(target);
-        if let Some(path) = node_path(&scene, target) {
-            st.selected_paths.push(path);
-        }
-        st.gizmo_drag = None;
-        st.pick_cycle = None;
-        true
-    }
-
-    /// Set the desired selection by stable paths. Live `NodeId`s are cleared
-    /// so the inspector doesn't render against stale indices; the next
-    /// `set_scene` call resolves the paths back to `NodeId`s once the
-    /// recompile lands.
-    pub fn set_selected_paths(&self, paths: Vec<SelectionPath>) {
-        let mut st = self.state.lock().unwrap();
-        st.selected_paths = paths;
-        st.selected.clear();
-        st.gizmo_drag = None;
-    }
-
-    pub fn gizmo_mode(&self) -> crate::gizmo::GizmoMode {
-        self.state.lock().unwrap().gizmo_mode
-    }
-
-    pub fn set_preview_shader(&self, shader: PreviewShader) {
-        self.state.lock().unwrap().preview_shader = shader;
-    }
-
-    pub fn set_show_grid(&self, on: bool) {
-        self.state.lock().unwrap().show_grid = on;
-    }
-
-    pub fn set_show_light_gizmos(&self, on: bool) {
-        self.state.lock().unwrap().show_light_gizmos = on;
-    }
-
-    pub fn set_show_transform_gizmo(&self, on: bool) {
-        self.state.lock().unwrap().show_transform_gizmo = on;
-    }
-
-    pub fn set_show_colliders(&self, on: bool) {
-        self.state.lock().unwrap().show_colliders = on;
-    }
-
-    /// Swap in a fresh environment-lighting preset. The next viewport paint
-    /// pulls `state.environment` and forwards its resolved params to the
-    /// renderer's sky-probe + key/fill uniforms.
-    pub fn set_environment(&self, env: Environment) {
-        self.state.lock().unwrap().environment = env;
-    }
-
-    pub fn environment(&self) -> Environment {
-        self.state.lock().unwrap().environment
-    }
-
-    /// Update the shadow-quality preset. The actual GPU-resource resize
-    /// happens lazily in the paint callback where a `glow::Context` is in
-    /// scope; this just stashes the desired quality. The renderer compares
-    /// against its own cached quality each paint and reallocates only when
-    /// they diverge.
-    pub fn set_shadows(&self, quality: shadows::ShadowQuality) {
-        self.state.lock().unwrap().shadows = quality;
-    }
-
-    pub fn shadows(&self) -> shadows::ShadowQuality {
-        self.state.lock().unwrap().shadows
-    }
-
-    /// Cap continuous viewport repaints (animation, cinema, gizmo drag).
-    /// `None` = uncapped. The cap is applied by routing the per-frame
-    /// repaint request through `request_repaint_after(1 / fps)` instead of
-    /// the immediate variant — input-driven paints still fire on demand.
-    pub fn set_max_fps(&self, max_fps: Option<u32>) {
-        self.state.lock().unwrap().max_fps = max_fps;
-    }
-
-    pub fn set_gizmo_mode(&self, mode: crate::gizmo::GizmoMode) {
-        let mut st = self.state.lock().unwrap();
-        st.gizmo_mode = mode;
-        st.gizmo_drag = None;
-    }
-
-    pub fn take_pending_edits(&self) -> Vec<PendingEdit> {
-        std::mem::take(&mut self.state.lock().unwrap().pending_edits)
-    }
-
-    pub fn take_pending_caret(&self) -> Option<usize> {
-        self.state.lock().unwrap().pending_caret.take()
-    }
-
-    pub fn push_pending_edit(&self, edit: PendingEdit) {
-        self.state.lock().unwrap().pending_edits.push(edit);
-    }
-
-    pub fn clips_snapshot(&self) -> Vec<ClipSummary> {
-        let st = self.state.lock().unwrap();
-        st.scene
-            .as_ref()
-            .map(|s| {
-                s.clips
-                    .iter()
-                    .map(|c| ClipSummary {
-                        name: c.name.clone(),
-                        duration: c.duration,
-                        origin: c.origin.clone(),
-                    })
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    pub fn active_clips(&self) -> Vec<bool> {
-        self.state.lock().unwrap().clip_active.clone()
-    }
-
-    pub fn set_clip_active(&self, idx: usize, active: bool) {
-        let mut st = self.state.lock().unwrap();
-        if idx >= st.clip_active.len() || st.clip_active[idx] == active {
-            return;
-        }
-        st.clip_active[idx] = active;
-        st.anim_times[idx] = 0.0;
-        st.update_palettes();
-    }
-
-    pub fn set_all_clips_active(&self, active: bool) {
-        let mut st = self.state.lock().unwrap();
-        let mut changed = false;
-        let n = st.clip_active.len();
-        for i in 0..n {
-            if st.clip_active[i] != active {
-                st.clip_active[i] = active;
-                st.anim_times[i] = 0.0;
-                changed = true;
-            }
-        }
-        if changed {
-            st.update_palettes();
-        }
-    }
-
-    pub fn is_playing(&self) -> bool {
-        self.state.lock().unwrap().anim_playing
-    }
-
-    pub fn set_playing(&self, playing: bool) {
-        self.state.lock().unwrap().anim_playing = playing;
-    }
-
-    pub fn playback_speed(&self) -> f32 {
-        self.state.lock().unwrap().playback_speed
-    }
-
-    pub fn set_playback_speed(&self, speed: f32) {
-        self.state.lock().unwrap().playback_speed = speed;
-    }
-
-    pub fn reset_anim_times(&self) {
-        let mut st = self.state.lock().unwrap();
-        for t in st.anim_times.iter_mut() {
-            *t = 0.0;
-        }
-        st.update_palettes();
-    }
-
-    pub fn anim_times(&self) -> Vec<f32> {
-        self.state.lock().unwrap().anim_times.clone()
-    }
-
-    pub fn seek_clip(&self, idx: usize, t: f32) {
-        let mut st = self.state.lock().unwrap();
-        if idx >= st.anim_times.len() {
-            return;
-        }
-        let duration = st
-            .scene
-            .as_ref()
-            .and_then(|s| s.clips.get(idx))
-            .map(|c| c.duration)
-            .unwrap_or(0.0);
-        let clamped = if duration > 0.0 {
-            t.rem_euclid(duration)
-        } else {
-            0.0
-        };
-        if (st.anim_times[idx] - clamped).abs() < 1e-5 {
-            return;
-        }
-        st.anim_times[idx] = clamped;
-        let active = *st.clip_active.get(idx).unwrap_or(&false);
-        if active {
-            st.update_palettes();
-        }
-    }
-
-    pub fn frame_view(&self) {
-        let mut st = self.state.lock().unwrap();
-        Self::frame_camera_to_static(&mut st);
-    }
-
-    fn frame_camera_to_static(st: &mut ViewerState) {
-        let center = st.static_center;
-        let radius = st.static_radius.max(0.001);
-        st.camera.target = center;
-        st.camera.fit_distance = radius * 2.8;
-        st.camera.zoom = 1.0;
-        st.camera.yaw = std::f32::consts::FRAC_PI_4;
-        st.camera.pitch = 0.5;
-    }
-
-    pub fn camera_snapshot(&self) -> CameraSnapshot {
-        self.state.lock().unwrap().camera.snapshot()
-    }
-
-    pub fn restore_camera(&self, snap: CameraSnapshot) {
-        self.state.lock().unwrap().camera.restore(snap);
-    }
-
-    pub fn is_cinema_active(&self) -> bool {
-        self.state.lock().unwrap().cinema.active
-    }
-
-    pub fn cinema_shot_label(&self) -> Option<&'static str> {
-        self.state.lock().unwrap().cinema.shot_label()
-    }
-
-    /// Toggle cinema mode. On enable, latches the current camera pose and
-    /// re-frames against the static bounding sphere so each shot composes
-    /// around the model itself. On disable, restores the latched pose.
-    ///
-    /// `force_play` controls whether animations should be force-started when
-    /// cinema activates. Pass `true` to match the previous default (the
-    /// model performs while the camera pans); `false` lets the user enable
-    /// cinema on a static model without surprise animation playback.
-    pub fn set_cinema_active(&self, on: bool, force_play: bool) {
-        let mut st = self.state.lock().unwrap();
-        if on == st.cinema.active {
-            return;
-        }
-        if on {
-            let center = st.static_center;
-            let radius = st.static_radius.max(0.001);
-            st.camera.target = center;
-            st.camera.fit_distance = radius * 2.8;
-            // Split the guard's deref into a single &mut ViewerState so the
-            // borrow checker can prove `cinema` and `camera` are disjoint.
-            let st = &mut *st;
-            st.cinema.activate(&st.camera);
-            st.gizmo_drag = None;
-            if force_play {
-                st.anim_playing = true;
-            }
-        } else if let Some(snap) = st.cinema.deactivate() {
-            st.camera.restore(snap);
-        }
-    }
-
     pub fn destroy(&self, gl: &glow::Context) {
         if let Ok(mut r) = self.renderer.lock() {
             r.destroy(gl);
         }
-    }
-
-    /// Queue an offscreen render. The next paint callback consumes the
-    /// request, renders each frame at `request.size × request.size` to a
-    /// fresh FBO, and writes a `CaptureOutcome` back into the viewer state.
-    /// The app drains it via [`Self::take_capture_outcome`] on subsequent
-    /// frames. Replaces any earlier in-flight request — the studio wires a
-    /// single capture-in-flight slot at the app level so this never collides
-    /// in practice.
-    pub fn submit_capture(&self, mut request: CaptureRequest) {
-        // Re-initialise progress bookkeeping every submission so callers
-        // don't have to know about `total` / `written` / `error` — they just
-        // hand us the work and we record the denominator here.
-        request.total = request.frames.len() as u32;
-        request.written = Vec::with_capacity(request.frames.len());
-        request.error = None;
-        let mut st = self.state.lock().unwrap();
-        st.capture_request = Some(request);
-        st.capture_outcome = None;
-    }
-
-    /// Drain the completed capture only when its kind matches `predicate`.
-    /// Used to keep `poll_generate` and the picker's background thumbnail
-    /// engine from racing over the single `capture_outcome` slot — each
-    /// caller passes a filter that matches only the kinds it owns. Without
-    /// this discrimination, whichever caller polls first (currently
-    /// `poll_generate`) drains every outcome, including the picker's, and
-    /// the picker's thumbnail pipeline stalls after one render.
-    pub fn take_capture_outcome_if(
-        &self,
-        predicate: impl FnOnce(CaptureKind) -> bool,
-    ) -> Option<CaptureOutcome> {
-        let mut st = self.state.lock().unwrap();
-        let matches = st
-            .capture_outcome
-            .as_ref()
-            .map(|o| predicate(o.kind))
-            .unwrap_or(false);
-        if matches {
-            st.capture_outcome.take()
-        } else {
-            None
-        }
-    }
-
-    /// Whether a capture request is queued or actively rendering. The app
-    /// uses this to gate menu items so the user can't pile up captures.
-    pub fn is_capturing(&self) -> bool {
-        self.state.lock().unwrap().capture_request.is_some()
-    }
-
-    /// Queue an imposter atlas bake. The next paint callback runs the bake
-    /// on the live GL context and writes the result to `imposter_outcome`.
-    /// Replaces any prior queued request — Studio gates this at the app
-    /// level so two callers never collide.
-    pub fn submit_imposter_request(&self, request: ImposterRequest) {
-        let mut st = self.state.lock().unwrap();
-        st.imposter_request = Some(request);
-        st.imposter_outcome = None;
-    }
-
-    /// Drain the most recent completed imposter bake, if any.
-    pub fn take_imposter_outcome(&self) -> Option<ImposterOutcome> {
-        self.state.lock().unwrap().imposter_outcome.take()
-    }
-
-    /// Whether an imposter request is queued or in-flight.
-    pub fn imposter_in_flight(&self) -> bool {
-        let st = self.state.lock().unwrap();
-        st.imposter_request.is_some() && st.imposter_outcome.is_none()
-    }
-
-    /// Enter or leave the viewport imposter-preview mode. When `scene` is
-    /// `Some`, the paint callback bakes (or re-bakes) the yaw-grid atlas
-    /// off it and renders a billboard quad in its place. When `scene` is
-    /// `None` (or `active` is false), the cached overlay texture is freed
-    /// next paint and the normal scene draw resumes.
-    pub fn set_imposter_view(&self, active: bool, scene: Option<Arc<SceneGraph>>) {
-        let mut st = self.state.lock().unwrap();
-        st.imposter_view_active = active;
-        if active {
-            // A fresh scene Arc invalidates any cached atlas — flip dirty
-            // so the next paint re-bakes against the latest geometry.
-            // Identity-compare the Arc so unrelated repaints (camera
-            // orbit, etc.) don't trigger a needless re-bake.
-            let same = match (&st.imposter_view_scene, &scene) {
-                (Some(prev), Some(next)) => Arc::ptr_eq(prev, next),
-                _ => false,
-            };
-            if !same {
-                st.imposter_view_dirty = true;
-                st.imposter_view_scene = scene;
-            }
-        } else {
-            // Leaving the mode — drop the scene reference; the texture is
-            // freed in the paint callback (needs &gl).
-            st.imposter_view_scene = None;
-            st.imposter_view_dirty = true;
-        }
-    }
-
-
-    /// Snapshot the in-flight capture's progress for the modal: the kind
-    /// (so the dialog can title itself "thumbnail" vs "video"), how many
-    /// frames have already been written, and the original frame count.
-    /// Returns `None` when no capture is queued.
-    pub fn capture_progress(&self) -> Option<(CaptureKind, u32, u32)> {
-        let st = self.state.lock().unwrap();
-        st.capture_request
-            .as_ref()
-            .map(|r| (r.kind, r.written.len() as u32, r.total))
     }
 
     pub fn show(&self, ui: &mut egui::Ui) -> egui::Response {
@@ -940,10 +402,10 @@ impl Viewer {
             // the previous frame's clock).
             rr.set_frame_time(frame_time);
             if st.capture_request.is_some() {
-                process_capture_step(&mut rr, gl, &mut st);
+                paint_jobs::process_capture_step(&mut rr, gl, &mut st);
             }
             if st.imposter_request.is_some() {
-                process_imposter_step(gl, &mut st);
+                paint_jobs::process_imposter_step(gl, &mut st);
             }
             // Free any cached billboard atlas when the user leaves
             // imposter view, and re-bake when entering / when the
@@ -954,7 +416,7 @@ impl Viewer {
                 }
                 st.imposter_view_dirty = false;
             } else if st.imposter_view_dirty {
-                process_imposter_view_bake(gl, &mut st);
+                paint_jobs::process_imposter_view_bake(gl, &mut st);
                 st.imposter_view_dirty = false;
             }
             let viewproj = st.camera.view_proj(aspect);
@@ -1078,261 +540,6 @@ impl Viewer {
 
         response
     }
-}
-
-
-/// Render one pending frame and hand the pixels to a background encoder.
-/// Runs inside the paint callback because that's where we have access to
-/// a `glow::Context`; the renderer's `render_to_pixels` already restores
-/// the bound FBO + viewport before returning, so this never leaks into
-/// egui's draw state. PNG encoding + disk I/O happen on the
-/// [`state::EncodePool`] worker threads so the GL thread can render the
-/// next frame as soon as `glReadPixels` returns instead of blocking on
-/// deflate.
-fn process_capture_step(
-    rr: &mut renderer::Renderer,
-    gl: &glow::Context,
-    st: &mut state::ViewerState,
-) {
-    // Phase 1: drain whatever the encoder pool has finished since last
-    // paint. Each completed encode either contributes a path to `written`
-    // or sets the request's first-fatal-error slot.
-    drain_encode_results(st);
-
-    // Phase 2: decide whether to finalise. We can only finalise once
-    // `frames` is drained AND there are no encodes still in flight —
-    // otherwise the outcome's `frame_paths` would be missing PNGs that
-    // workers are still writing.
-    let in_flight = st.encode_pool.as_ref().map(|p| p.in_flight).unwrap_or(0);
-    let (frames_done, errored) = match st.capture_request.as_ref() {
-        Some(req) => (req.frames.is_empty(), req.error.is_some()),
-        None => return,
-    };
-    if (frames_done || errored) && in_flight == 0 {
-        // Drop the pool first: workers exit as soon as `job_tx` closes,
-        // and dropping here means a fresh capture starts with a fresh
-        // pool rather than reusing one that's already been signalled.
-        st.encode_pool = None;
-        let req = st.capture_request.take().expect("checked above");
-        st.capture_outcome = Some(CaptureOutcome {
-            kind: req.kind,
-            frame_paths: req.written,
-            error: req.error,
-        });
-        return;
-    }
-    if errored {
-        // Don't queue any more renders once a fatal error is recorded;
-        // just keep paint cycles ticking so phase 1 can drain whatever
-        // encodes were already in flight when the error fired.
-        return;
-    }
-    if frames_done {
-        // Frames all submitted but encodes still pending — nothing to do
-        // on the GL side this paint, just wait for the pool to catch up.
-        return;
-    }
-
-    // Phase 3: render the next frame. The borrow scoping here mirrors the
-    // pre-async version: pull the per-frame inputs out of `req` in a
-    // narrow scope so we can call `st.update_palettes()` later without
-    // holding two mutable borrows.
-    let (size, bg, kind, frame) = {
-        let req = st.capture_request.as_mut().expect("checked above");
-        let f = req.frames.remove(0);
-        (req.size, req.bg, req.kind, f)
-    };
-
-    let center = st.static_center;
-    // Floor on the framing radius so a one-vertex / empty scene still picks
-    // a sane orbit distance — without this, `radius * 2.8` collapses to 0
-    // and the camera ends up inside the model.
-    let radius = st.static_radius.max(0.001);
-    let cam = camera::OrbitCamera {
-        yaw: frame.yaw,
-        pitch: frame.pitch,
-        fit_distance: radius * 2.8,
-        zoom: 1.0,
-        target: center,
-    };
-    let viewproj = cam.view_proj(1.0);
-    let eye = cam.eye();
-    let frame_time = frame.time;
-    // Video frames want the animation sampled at `frame.time` so the encoded
-    // mp4 plays clips back across the rotation. Thumbnails ignore time and
-    // capture whatever pose is currently visible.
-    let anim_override = kind == CaptureKind::Video
-        && st.any_active()
-        && st
-            .scene
-            .as_ref()
-            .map(|s| !s.clips.is_empty())
-            .unwrap_or(false);
-    let saved_anim_times = if anim_override {
-        let snapshot = st.anim_times.clone();
-        // Collect durations up front so we can index `st.anim_times` mutably
-        // without holding a `&Scene` borrow across the loop.
-        let durations: Vec<f32> = st
-            .scene
-            .as_ref()
-            .map(|s| s.clips.iter().map(|c| c.duration).collect())
-            .unwrap_or_default();
-        for i in 0..st.clip_active.len() {
-            if !st.clip_active[i] {
-                continue;
-            }
-            let duration = durations.get(i).copied().unwrap_or(0.0);
-            if duration > 0.0 {
-                st.anim_times[i] = frame_time.rem_euclid(duration);
-            }
-        }
-        st.update_palettes();
-        rr.upload_palettes(&st.mesh.palettes);
-        st.palettes_dirty = false;
-        Some(snapshot)
-    } else {
-        None
-    };
-    let render_result = rr.render_to_pixels(gl, size, viewproj, eye, bg);
-    // Restore palettes before touching anything else so an on-screen draw
-    // that follows in this same paint callback matches the user's pose.
-    if let Some(snapshot) = saved_anim_times {
-        st.anim_times = snapshot;
-        st.update_palettes();
-        rr.upload_palettes(&st.mesh.palettes);
-        st.palettes_dirty = false;
-    }
-    match render_result {
-        Ok(pixels) => {
-            // Lazy-init the pool so a never-captured studio session never
-            // pays for the worker threads.
-            let pool = st
-                .encode_pool
-                .get_or_insert_with(state::EncodePool::new);
-            pool.submit(pixels, size, frame.path);
-        }
-        Err(e) => {
-            if let Some(req) = st.capture_request.as_mut() {
-                req.error = Some(format!("render: {e}"));
-            }
-        }
-    }
-}
-
-/// Drain everything the encoder pool has produced since last paint into
-/// the live `CaptureRequest`. Successful encodes append to `written`;
-/// the first failure latches into `error` and short-circuits future
-/// frames in the next `process_capture_step` call.
-fn drain_encode_results(st: &mut state::ViewerState) {
-    let Some(pool) = st.encode_pool.as_mut() else {
-        return;
-    };
-    while let Ok((path, res)) = pool.result_rx.try_recv() {
-        // Underflow guard: in-flight should always match the number of
-        // outstanding sends, but a stray send from a dropped-and-recreated
-        // pool would otherwise wrap to usize::MAX and freeze finalisation.
-        pool.in_flight = pool.in_flight.saturating_sub(1);
-        let Some(req) = st.capture_request.as_mut() else {
-            continue;
-        };
-        match res {
-            Ok(()) => req.written.push(path),
-            Err(e) => {
-                if req.error.is_none() {
-                    req.error = Some(e);
-                }
-            }
-        }
-    }
-}
-
-/// Re-bake the viewport imposter atlas and upload it as a GL texture.
-/// Called from the paint callback when `imposter_view_dirty` is set on
-/// entering imposter mode or after a scene recompile. Frees any prior
-/// texture before uploading the new one.
-///
-/// Bake parameters mirror the export's defaults (256² cells, 8 yaws,
-/// 0.5 rad pitch) so the in-Studio preview shows the same artifact the
-/// shipped GLB embeds.
-fn process_imposter_view_bake(gl: &glow::Context, st: &mut state::ViewerState) {
-    // Free the prior texture first so a bake failure still frees the GL
-    // resource instead of leaking it on every failed re-bake.
-    if let Some(prev) = st.imposter_view_overlay.take() {
-        renderer::Renderer::destroy_imposter_texture(gl, prev.texture);
-    }
-    let Some(scene) = st.imposter_view_scene.as_ref() else {
-        return;
-    };
-    // Match the export's bake parameters so what the user previews
-    // here is exactly what `bundle_lods_and_imposter` would ship —
-    // same cell size, view count, and pitch. The atlas returns the
-    // pitch-aware quad placement so we don't need to compute extents
-    // separately on this side.
-    //
-    // Pass through the viewer's `base_dir` so `flatten` can resolve
-    // relative material texture paths into absolute ones — without
-    // this the bake's per-cell render falls back to PBR scalars and
-    // every silhouette comes out flat-coloured (no baseColor /
-    // normal / roughness textures applied).
-    let opts = mogen_render::imposter::ImposterOptions {
-        cell_size: 512,
-        view_count: 8,
-        pitch: 0.5,
-        base_dir: st.base_dir.clone(),
-    };
-    let atlas = match mogen_render::imposter::bake_yaw_atlas_on_gl(gl, scene, &opts) {
-        Ok(a) => a,
-        Err(_e) => {
-            // Bake failures here are silent — the billboard simply
-            // doesn't draw and the user gets the grid + background.
-            // Surfacing the error in the viewport would need an overlay
-            // layer we don't have yet; the imposter preview modal
-            // already shows bake errors when the user opens it.
-            return;
-        }
-    };
-    let texture =
-        match renderer::Renderer::upload_imposter_atlas(gl, &atlas.rgba, atlas.width, atlas.height)
-        {
-            Ok(t) => t,
-            Err(_) => return,
-        };
-    st.imposter_view_overlay = Some(state::ImposterViewOverlay {
-        texture,
-        view_count: atlas.view_count,
-        center: atlas.center,
-        half_width: atlas.half_width,
-        half_height: atlas.half_height,
-        uv_y_top: atlas.uv_y_top,
-        uv_y_bottom: atlas.uv_y_bottom,
-    });
-}
-
-/// Service a queued imposter atlas bake on the live `glow::Context`. Runs
-/// inside the paint callback, alongside `process_capture_step`, because
-/// `mogen_render::imposter::bake_yaw_atlas_on_gl` allocates its own
-/// `Renderer` and FBOs on the GL context — eframe owns the only winit
-/// `EventLoop` in the process, so the CLI's `with_gl_context` path can't
-/// be used here.
-///
-/// The bake runs end-to-end in one paint (8 yaws × 256² each is well under
-/// a frame budget). Result lands in `imposter_outcome` for the app to poll.
-fn process_imposter_step(gl: &glow::Context, st: &mut state::ViewerState) {
-    let Some(request) = st.imposter_request.take() else {
-        return;
-    };
-    let outcome = mogen_render::imposter::bake_yaw_atlas_on_gl(
-        gl,
-        &request.scene,
-        &mogen_render::imposter::ImposterOptions {
-            cell_size: request.cell_size,
-            view_count: request.view_count,
-            pitch: request.pitch,
-            base_dir: request.base_dir.clone(),
-        },
-    )
-    .map_err(|e| format!("{e:#}"));
-    st.imposter_outcome = Some(outcome);
 }
 
 #[cfg(test)]
